❖ | sᴧᴄʜɪɴ ꭙ sᴧɴᴧᴛᴧɴɪ | ❖

+++ SACHIN_MUSIC/plugins/tools/telegraph.py
@@ -1,8 +1,20 @@
 import os
 from pyrogram import filters
-from pyrogram.types import InlineKeyboardMarkup, InlineKeyboardButton
-from telegraph import upload_file
+from pyrogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 from SACHIN_MUSIC import app
+import requests
+
+
+def upload_file(file_path):
+    url = "https://catbox.moe/user/api.php"
+    data = {"reqtype": "fileupload", "json": "true"}
+    files = {"fileToUpload": open(file_path, "rb")}
+    response = requests.post(url, data=data, files=files)
+
+    if response.status_code == 200:
+        return True, response.text.strip()
+    else:
+        return False, f"❖ ᴇʀʀᴏʀ : {response.status_code} - {response.text}"
 
 
 @app.on_message(filters.command(["tgm", "tgt", "telegraph", "tl"]))
@@ -11,36 +23,64 @@ async def get_link_group(client, message):
         return await message.reply_text(
             "❖ ᴘʟᴇᴀsᴇ ʀᴇᴘʟʏ ᴛᴏ ᴀ ᴍᴇᴅɪᴀ ᴛᴏ ᴜᴘʟᴏᴀᴅ ᴏɴ ᴛᴇʟᴇɢʀᴀᴘʜ"
         )
+
+    media = message.reply_to_message
+    file_size = 0
+    if media.photo:
+        file_size = media.photo.file_size
+    elif media.video:
+        file_size = media.video.file_size
+    elif media.document:
+        file_size = media.document.file_size
+
+    if file_size > 200 * 1024 * 1024:
+        return await message.reply_text("ᴘʟᴇᴀsᴇ ᴘʀᴏᴠɪᴅᴇ ᴀ ᴍᴇᴅɪᴀ ғɪʟᴇ ᴜɴᴅᴇʀ 200 MB")
+
     try:
         text = await message.reply("❍ ᴘʀᴏᴄᴇssɪɴɢ...")
 
         async def progress(current, total):
-            await text.edit_text(f"❍ ᴅᴏᴡɴʟᴏᴀᴅɪɴɢ... {current * 100 / total:.1f}%")
+            try:
+                await text.edit_text(f"❍ ᴅᴏᴡɴʟᴏᴀᴅɪɴɢ... {current * 100 / total:.1f}%")
+            except Exception:
+                pass
 
         try:
-            location = f"cache"
-            local_path = await message.reply_to_message.download(
-                location, progress=progress
-            )
+            local_path = await media.download(progress=progress)
             await text.edit_text("❍ ᴜᴘʟᴏᴀᴅɪɴɢ ᴛᴏ ᴛᴇʟᴇɢʀᴀᴘʜ...")
-            upload_path = upload_file(local_path)
-            await text.edit_text(
-                f"❖ | [ᴛᴇʟᴇɢʀᴀᴘʜ ʟɪɴᴋ](https://telegra.ph{upload_path[0]}) | ❖",
-                reply_markup=InlineKeyboardMarkup(
-                    [
+
+            success, upload_path = upload_file(local_path)
+
+            if success:
+                await text.edit_text(
+                    f"❖ | [ᴛᴇʟᴇɢʀᴀᴘʜ ʟɪɴᴋ]({upload_path}) | ❖",
+                    reply_markup=InlineKeyboardMarkup(
                         [
-                            InlineKeyboardButton(
-                                "• ᴛᴇʟᴇɢʀᴀᴘʜ ʟɪɴᴋ •",
-                                url=f"https://telegra.ph{upload_path[0]}",
-                            )
+                            [
+                                InlineKeyboardButton(
+                                    "• ᴛᴇʟᴇɢʀᴀᴘʜ ʟɪɴᴋ •",
+                                    url=upload_path,
+                                )
+                            ]
                         ]
-                    ]
-                ),
-            )
-            os.remove(local_path)
+                    ),
+                )
+            else:
+                await text.edit_text(
+                    f"❖ ᴀɴ ᴇʀʀᴏʀ ᴏᴄᴄᴜʀʀᴇᴅ ᴡʜɪʟᴇ ᴜᴘʟᴏᴀᴅɪɴɢ ʏᴏᴜʀ ғɪʟᴇ\n{upload_path}"
+                )
+
+            try:
+                os.remove(local_path)
+            except Exception:
+                pass
+
         except Exception as e:
-            await text.edit_text(f"❍ |ғɪʟᴇ ᴜᴘʟᴏᴀᴅ ғᴀɪʟᴇᴅ \n\n<i>ʀᴇᴀsᴏɴ: {e}</i>")
-            os.remove(local_path)
+            await text.edit_text(f"❖ | ғɪʟᴇ ᴜᴘʟᴏᴀᴅ ғᴀɪʟᴇᴅ\n\n<i>❍ ʀᴇᴀsᴏɴ : {e}</i>")
+            try:
+                os.remove(local_path)
+            except Exception:
+                pass
             return
     except Exception:
         pass
